requirements.txt
azure-servicebus==7.11.1
